Examples of day 1 fixed
