+++ src/main/java/edu/solid/Flyer.java
@@ -1,0 +1,5 @@
+package edu.solid;
+
+public interface Flyer{
+    void fly();
+}
