+++ src/main/java/edu/solid/Quacker.java
@@ -1,0 +1,5 @@
+package edu.solid;
+
+public interface Quacker{
+    void quack();
+}
